feat: Refactor API
