+++ src/components/AdminDashboard/PendingApplicationDashboard/AcceptPendingApplicationButton.tsx
@@ -5,7 +5,7 @@ import Button from "@mui/material/Button";
 import React, { Dispatch, SetStateAction } from "react";
 
 import { Row } from "@/components/AdminDashboard/PendingApplicationDashboard";
-import { handleApproveProgramApplication } from "@/server/api/program-enrollments/mutations";
+import { handleApproveProgramApplication } from "@/server/api/program-enrollments/public-mutations";
 import { Program } from "@/types";
 
 type AcceptPendingApplicationButtonProps = {
