+++ src/server/api/password-reset-tokens/private-mutations.ts
@@ -1,13 +1,8 @@
-"use server";
-
-import { sendPasswordResetEmail } from "@/server/api/emails/actions";
 import { getPasswordResetTokenByToken } from "@/server/api/password-reset-tokens/queries";
-import { getUserByEmail } from "@/server/api/users/queries";
 import dbConnect from "@/server/dbConnect";
 import { PasswordResetTokenModel } from "@/server/models";
 import { ApiResponse, PasswordResetToken } from "@/types";
 import apiErrors from "@/utils/constants/apiErrors";
-import dayjsUtil from "@/utils/dayjsUtil";
 import handleMongooseError from "@/utils/handleMongooseError";
 
 export async function createPasswordResetToken(
@@ -67,37 +62,4 @@ export async function deletePasswordResetToken(
   } catch (error) {
     return [null, handleMongooseError(error)];
   }
-}
-
-export async function handlePasswordResetRequest(email: string) {
-  // check if email exists
-  const [user, userError] = await getUserByEmail(email);
-
-  if (userError !== null) {
-    return;
-  }
-
-  if (!user._id) {
-    return;
-  }
-
-  // generate a password reset token
-  const oneHourFromNow = dayjsUtil().utc().add(1, "hour").toISOString();
-  const token: PasswordResetToken = {
-    token: crypto.randomUUID(),
-    userId: user._id,
-    expires: oneHourFromNow,
-  };
-
-  const [, createPasswordResetTokenError] = await createPasswordResetToken(
-    token.token,
-    token.userId,
-    token.expires,
-  );
-
-  if (createPasswordResetTokenError !== null) {
-    return;
-  }
-
-  await sendPasswordResetEmail(email, token.token);
 }
